+++ encoder_config_test.go
@@ -38,10 +38,9 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 		expected string
 	}{
 		{name: "defaultConfig",
-			cfg:   NewDefaultEncoderConfig(),
-			input: `{"timeEncoder":"millis"}`,
+			cfg: NewDefaultEncoderConfig(),
 			expected: `{"log.level": "debug",
-						"@timestamp": 1583484083953.468,
+						"@timestamp": "2020-09-13T10:48:03.000Z",
 						"message": "log message",
 						"log.origin": {
 							"file.line": 30,
@@ -52,9 +51,8 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 						"foo": "bar",
 						"dur": 5000000}`},
 		{name: "defaultUnmarshal",
-			input: `{"timeEncoder":"millis"}`,
 			expected: `{"log.level": "debug",
-						"@timestamp": 1583484083953.468,
+						"@timestamp": "2020-09-13T10:48:03.000Z",
 						"message": "log message",
 						"foo": "bar",
 						"dur": 5000000}`},
@@ -63,11 +61,10 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
   					 "enableStackTrace": true,
 					 "enableCaller":true,
 					 "levelEncoder": "upper",
-					 "timeEncoder":"nanos",
 				 	 "durationEncoder": "ms",
 					 "callerEncoder": "short"}`,
 			expected: `{"log.level": "debug",
-						"@timestamp": 1583484083953467800,
+						"@timestamp": "2020-09-13T10:48:03.000Z",
 						"message": "log message",
 						"log.origin": {
 							"file.line": 30,
@@ -78,9 +75,9 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 						"foo": "bar",
 						"dur": 5}`},
 		{name: "fullCaller",
-			input: `{"callerEncoder": "full","enableCaller":true,"timeEncoder":"millis"}`,
+			input: `{"callerEncoder": "full","enableCaller":true}`,
 			expected: `{"log.level": "debug",
-						"@timestamp": 1583484083953.468,
+						"@timestamp": "2020-09-13T10:48:03.000Z",
 						"message": "log message",
 						"log.origin": {
 							"file.line": 30,
@@ -93,7 +90,7 @@ func TestJSONEncoder_EncoderConfig(t *testing.T) {
 			// setup
 			entry := zapcore.Entry{
 				Level:      zapcore.DebugLevel,
-				Time:       time.Unix(1583484083, 953467845),
+				Time:       time.Unix(1599994083, 0).UTC(),
 				Message:    "log message",
 				Caller:     caller,
 				Stack:      "frames",
@@ -127,9 +124,9 @@ func TestECSCompatibleEncoderConfig(t *testing.T) {
 		expected string
 	}{
 		{name: "empty config",
-			cfg: zapcore.EncoderConfig{EncodeTime: zapcore.EpochTimeEncoder},
+			cfg: zapcore.EncoderConfig{},
 			expected: `{"log.level": "debug",
-						"@timestamp": 1583484083.9534678,
+						"@timestamp": "2020-09-13T10:48:03.000Z",
 						"message": "log message",
 						"foo": "bar",
 						"count": 8}`},
@@ -140,7 +137,7 @@ func TestECSCompatibleEncoderConfig(t *testing.T) {
 				NameKey: "replaced nameKey", StacktraceKey: "replaced stackTraceKey",
 				CallerKey: "replaced callerKey", EncodeLevel: zapcore.CapitalLevelEncoder},
 			expected: `{"log.level": "DEBUG",
-						"@timestamp": 1583484083953.468,
+						"@timestamp": "2020-09-13T10:48:03.000Z",
 						"message": "log message",
 						"log.origin": {
 							"file.line": 30,
@@ -155,7 +152,7 @@ func TestECSCompatibleEncoderConfig(t *testing.T) {
 			// setup
 			entry := zapcore.Entry{
 				Level:      zapcore.DebugLevel,
-				Time:       time.Unix(1583484083, 953467845).UTC(),
+				Time:       time.Unix(1599994083, 0).UTC(),
 				Message:    "log message",
 				Caller:     caller,
 				Stack:      "frames",