+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for ecszap
 
 ### Bug Fixes
 * Change `stacktrace` to `stack_trace` in output and in json and yaml config option for `EncoderConfig.EnableStacktrace` [pull#21](https://github.com/elastic/ecs-logging-go-zap/pull/21)
+* Do not allow configuration of `@timestamp`, instead always set format to ISO8601 [pull#23](https://github.com/elastic/ecs-logging-go-zap/pull/23)
 
 ## 0.3.0